+++ src/main/java/kitchenpos/application/TableService.java
@@ -12,7 +12,7 @@ import kitchenpos.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.ui.dto.OrderTableRequest;
+import kitchenpos.ui.dto.request.OrderTableRequest;
 import kitchenpos.ui.dto.OrderTableResponse;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
