+++ src/main/java/kitchenpos/ui/TableRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.ui;
 import java.net.URI;
 import java.util.List;
 import kitchenpos.application.TableService;
-import kitchenpos.ui.dto.OrderTableNumberOfGuestsRequest;
-import kitchenpos.ui.dto.OrderTableRequest;
+import kitchenpos.ui.dto.request.OrderTableNumberOfGuestsRequest;
+import kitchenpos.ui.dto.request.OrderTableRequest;
 import kitchenpos.ui.dto.OrderTableResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
