+++ src/main/java/kitchenpos/ui/dto/request/MenuGroupRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.request;
 
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
