+++ src/main/java/kitchenpos/ui/dto/request/OrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.request;
 
 import kitchenpos.domain.OrderLineItem;
