+++ src/main/java/kitchenpos/ui/dto/request/OrderTableEmptyRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.request;
 
 public class OrderTableEmptyRequest {
     private boolean empty;
