+++ src/main/java/kitchenpos/ui/dto/request/OrderTableNumberOfGuestsRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.request;
 
 public class OrderTableNumberOfGuestsRequest {
     private int numberOfGuests;
