+++ src/test/java/kitchenpos/ui/TableRestControllerTest.java
@@ -15,9 +15,9 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.domain.OrderTable;
-import kitchenpos.ui.dto.OrderTableEmptyRequest;
-import kitchenpos.ui.dto.OrderTableNumberOfGuestsRequest;
-import kitchenpos.ui.dto.OrderTableRequest;
+import kitchenpos.ui.dto.request.OrderTableEmptyRequest;
+import kitchenpos.ui.dto.request.OrderTableNumberOfGuestsRequest;
+import kitchenpos.ui.dto.request.OrderTableRequest;
 import kitchenpos.ui.dto.OrderTableResponse;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
